+++ src/constants/notices.ts
@@ -1,0 +1,11 @@
+export default {
+  /**
+   * Notices to ignore.
+   */
+  IgnoredNotices: ['ban_success', 'timeout_success'],
+
+  /**
+   * Notices indicating that a user can be marked as unbanned.
+   */
+  UnbanNotices: ['bad_unban_no_ban', 'unban_success', 'untimeout_success'],
+}