+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 🚀 New Feature
 
 - Add a tool to quickly create a [Straw Poll](https://www.strawpoll.me/) (can be disabled in the settings).
-- Add possibility to unblock previously blocked users.
+- Add a button & context menu item to unban previously banned users.
+- Add a button to unblock previously blocked users.
 
 ### 🐛 Bug Fix